Make tagging nav hyperlink system for posts
I finished up all the tagging and autoloading posts for this branch finally. I
am replacing more and more of my md2html.sh script with properties from the
scruct itself with templating. I used the range templating function to add all
the links in as well as used other parts of the struct like the .Title and
.FileName members.