Translate URLs with translateURL
I made a ton of modifications to my templating system as well as my handlers to
get multilingual urls for real globalization on my site. example.org/tags/tag1
is still not working because of an issue with translating from nonenglish with
my funcs. Everything should be fixed soon though.