Design "working" rss feed
I make an atom feed with encoding/xml, and I'm probably gonna undo that in a
commit or two. I am having issues with self closing elements, and it seems the
only way to fix that is to get old libraries or steal scary funcs off of stack
overflow. I'll just write it with strings and loops tomorrow. I also removed
rss links from my tags to make posts seem more "official." Not sure if I'm
gonna keep that.