Plan for method based sql transactions
This is very cryptic sounding, but having a single package var for the path to
the database makes this useless for any other go app. I'm gonna move to using a
struct with methods so I don't have to supply the path in each time, but I get
the flexibility to extend this package to my cli app that I'll eventually make.