Append .html names to rss feed links
Gonna need to add dates to a lot of post links too. Making methods to
convert dates would make a ton of sense.

This is getting way too complicated though. I might do something crazy
and attempt to make a more flat template system for the sake of my
sanity.